Proposal optimizations + repo cleanup (#174)
* Attempt brute force changing of proposals and remove old contracts

* Remove anchor and anchor tests

* Formatting

* Spacing to tabs

* Cleanup unused files

* Test updates

* Cmt fmt

* Ups

* Fix all non-anchor proposals

* Fix tests except for native transactWrap

* Fix rescue tests

* Only call getAmountToWrap if the transact call is actually wrapping

* Fix path for contracts import

* More cleanup

* Fix error which allows users to prove against arbitrary edges if the edgeList is not at maxEdges length (#175)

* Compile

* Remove unnecessary ts-ignores

* Add file extension to SetupTxVanchorMock

* Fix capitalization

+++ contracts/interfaces/ITreasury.sol
@@ -17,12 +17,12 @@ interface ITreasury {
 		@param _amountToRescue Amount of tokens to rescue.
 		@param _nonce Nonce of the rescue transaction.
 	 */
-	function rescueTokens(address _tokenAddress, address payable _to, uint256 _amountToRescue, uint256 _nonce) external;
+	function rescueTokens(address _tokenAddress, address payable _to, uint256 _amountToRescue, uint32 _nonce) external;
 
 	/**
 		@notice Sets the handler responsible with relaying rescue transactions.
 		@param _newHandler Address of the handler.
 		@param _nonce Nonce of the update.
 	 */
-	function setHandler(address _newHandler, uint256 _nonce) external;
+	function setHandler(address _newHandler, uint32 _nonce) external;
 }
